% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{aq_pollutants_bcn}
\alias{aq_pollutants_bcn}
\title{Pollutants measured at air quality measurement stations of Barcelona.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.
}
\source{
BCN Open Data \url{https://opendata-ajuntament.barcelona.cat/data/en/dataset/contaminants-estacions-mesura-qualitat-aire}
}
\usage{
aq_pollutants_bcn
}
\description{
A list of pollutants measured at air quality measurement stations of Barcelona.

\describe{
\item{codi_contaminant}{Code of pollutant.}
\item{desc_contaminant}{Name of pollutant.}
\item{unitats}{Measurement units of pollutant.}
}
}
\examples{

aq_pollutants_bcn

}
\keyword{datasets}
