% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{aq_stations_bcn}
\alias{aq_stations_bcn}
\title{Air quality measurement stations of Barcelona.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 7 columns.
}
\source{
BCN Open Data \url{https://opendata-ajuntament.barcelona.cat/data/en/dataset/metadades-estacions-meteorologiques}
}
\usage{
aq_stations_bcn
}
\description{
A list of air quality measurement stations of Barcelona.

\describe{
\item{estacio}{Code of air quality measurement station.}
\item{nom_cabina}{Name of air quality measurement station.}
\item{longitud}{Longitude of air quality measurement station location.}
\item{latitud}{Latitude of air quality measurement station location.}
\item{codi_districte}{Code of district of air quality measurement station location.}
\item{codi_barri}{Code of neighbourhood of air quality measurement station location.}
\item{geometry}{Location of air quality measurement station location as sf object.}
}
}
\examples{

aq_stations_bcn

}
\keyword{datasets}
