% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{BCNNeigh}
\alias{BCNNeigh}
\title{A map of barcelona neighbourhoods}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 73 rows and 16 columns.
}
\source{
\url{https://vangdata.carto.com/tables/shapefiles_barcelona_distrito/public}
}
\usage{
BCNNeigh
}
\description{
A map of barcelona neighbourhoods loaded as a sf object

Neighbourhoods and district names have been set into utf8 encoding.
}
\examples{

BCNNeigh

}
\keyword{datasets}
