% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{cap_provincia_es}
\alias{cap_provincia_es}
\title{Provincial capitals of Spain.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52 rows and 9 columns.
}
\source{
Filtered from Business Intelligence \url{https://www.businessintelligence.info/resources/assets/listado-longitud-latitud-municipios-espana.xls}
}
\usage{
cap_provincia_es
}
\description{
A data frame of the provincial capitals of Spain.

\describe{
\item{comunidad}{Autonomous comunity.}
\item{provincia}{Province.}
\item{poblacion}{City/town name.}
\item{latitud}{Latitude.}
\item{longitud}{Longitude.}
\item{altitud}{Altitude.}
\item{habitantes}{Total population.}
\item{hombres}{Total population (men).}
\item{mujeres}{Total population (women).}
}
}
\examples{

cap_provincia_es

}
\keyword{datasets}
