% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{csv_airports}
\alias{csv_airports}
\title{A database of airports.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11776 rows and 12 columns.
}
\source{
unknown (didn't bother to record at the moment).
}
\usage{
csv_airports
}
\description{
Another database of airports. It comes from a .csv file of unknown origin, with some additions and modifications with information from Wikipedia.

\describe{
\item{name}{Name of airport.}
\item{city}{Main city served by airport.}
\item{country}{Country or territory where airport is located, in ISO 3166-1 format.}
\item{region}{State or administrative division of the country where the airport is located.}
\item{IATA}{3-letter IATA code.}
\item{ICAO}{4-letter ICAO code.}
\item{lat}{Decimal degrees, usually to six significant digits. Negative is South, positive is North.}
\item{lon}{Decimal degrees, usually to six significant digits. Negative is West, positive is East.}
\item{alt}{Altitude in feet.}
\item{timezone}{Hours offset from UTC. Fractional hours are expressed as decimals, eg. India is 5.5.}
\item{tz}{Timezone in "tz" (Olson) format, eg. "America/Los_Angeles".}
\item{WAC}{Numeric codes used to identify geopolitical areas such as countries, states, provinces, and territories or possessions of certain countries. The codes are used within the various data banks maintained by the Office of Airline Information (OAI) and are created by OAI.}
}
}
\examples{

csv_airports

}
\keyword{datasets}
