% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{esp_prov}
\alias{esp_prov}
\title{A Map of Spanish Provinces}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 52 rows and 24 columns.
}
\source{
Hernangómez D (2026). mapSpain: Administrative Boundaries of Spain. doi:10.5281/zenodo.5366622, \url{https://ropenspain.github.io/mapSpain/}.
}
\usage{
esp_prov
}
\description{
A map of the provinces of Spain, as obtained with mapSpain::esp_get_prov_siane(epsg = "4326")
}
\examples{

esp_prov

}
\keyword{datasets}
