% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{mwgg_airports}
\alias{mwgg_airports}
\title{The mwgg airport database.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 28870 rows and 10 columns.
}
\source{
\url{https://github.com/mwgg/Airports}
}
\usage{
mwgg_airports
}
\description{
Obtained from a JSON database of 28828 entries with basic information about nearly every airport and landing strip in the world. ICAO codes used as primary value. Some IATA codes have been added from the csv_airports database.

\describe{
\item{icao}{4-letter ICAO code.}
\item{iata}{3-letter IATA code.}
\item{name}{Name of airport.}
\item{city}{Main city served by airport.}
\item{country}{Country or territory where airport is located, in ISO 3166-1 format.}
\item{state}{State or administrative division of the country where the airport is located.}
\item{elevation}{Altitude in feet.}
\item{lat}{Decimal degrees, usually to ten significant digits. Negative is South, positive is North.}
\item{lon}{Decimal degrees, usually to ten significant digits. Negative is West, positive is East.}
\item{tz}{Timezone in "tz" (Olson) format, eg. "America/Los_Angeles".}
}
}
\examples{

mwgg_airports

}
\keyword{datasets}
