% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{of_routes}
\alias{of_routes}
\title{OpenFlights route database.}
\format{
An object of class \code{data.frame} with 67663 rows and 6 columns.
}
\source{
\url{https://openflights.org/data.html}
}
\usage{
of_routes
}
\description{
The set of OpenFlights routes. The third-party that OpenFlights uses for route data ceased providing updates in June 2014. The current data is of historical value only.

\describe{
\item{airline}{2-letter (IATA) or 3-letter (ICAO) code of the airline.}
\item{org}{3-letter (IATA) or 4-letter (ICAO) code of the source (origin) airport.}
\item{dst}{3-letter (IATA) or 4-letter (ICAO) code of the destination airport.}
\item{codeshare}{"Y" if this flight is a codeshare (that is, not operated by Airline, but another carrier), "N otherwise.}
\item{stops}{Number of stops on this flight ("0" for direct).}
\item{equipment}{3-letter codes for plane type(s) generally used on this flight, separated by spaces.}
}
}
\examples{

of_routes

}
\keyword{datasets}
