% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{poblaciones_es}
\alias{poblaciones_es}
\title{Spanish towns and cities.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8112 rows and 9 columns.
}
\source{
Business Intelligence \url{https://www.businessintelligence.info/resources/assets/listado-longitud-latitud-municipios-espana.xls}
}
\usage{
poblaciones_es
}
\description{
A data frame of all the towns and cities of Spain.

\describe{
\item{comunidad}{Autonomous comunity.}
\item{provincia}{Province.}
\item{poblacion}{City/town name.}
\item{latitud}{Latitude.}
\item{longitud}{Longitude.}
\item{altitud}{Altitude.}
\item{habitantes}{Total population.}
\item{hombres}{Total population (men).}
\item{mujeres}{Total population (women).}
}
}
\examples{

poblaciones_es

}
\keyword{datasets}
