% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{trs_barris}
\alias{trs_barris}
\title{A map of Terrassa neighbourhoods.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 56 rows and 8 columns.
}
\source{
\url{https://opendata.terrassa.cat/es/dataset/mapa-de-la-ciutat-amb-les-divisions-per-barris}
}
\usage{
trs_barris
}
\description{
A map of Terrassa neighbourhoods loaded as a sf object

Additional information: AREA and PERIMETER of each neighbourhood, and neighbourhood name in NOM_BARRI.
}
\examples{

trs_barris

}
\keyword{datasets}
