% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{trs_codis}
\alias{trs_codis}
\title{A map of Terrassa zip codes.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 6 columns.
}
\source{
\url{https://opendata.terrassa.cat/es/dataset/mapa-de-la-ciutat-amb-les-divisions-per-districtes}
}
\usage{
trs_codis
}
\description{
A map of Terrassa zip codes loaded as a sf object

Additional information: AREA and PERIMETER of each zip code, and zip code in CODI_POSTA.
}
\examples{

trs_codis

}
\keyword{datasets}
