% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{trs_districtes}
\alias{trs_districtes}
\title{A map of Terrassa districts.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7 rows and 6 columns.
}
\source{
\url{https://opendata.terrassa.cat/es/dataset/mapa-de-la-ciutat-amb-les-divisions-per-districtes}
}
\usage{
trs_districtes
}
\description{
A map of Terrassa districts loaded as a sf object

Additional information: AREA and PERIMETER of each district, and district number in DISTRICTE.
}
\examples{

trs_districtes

}
\keyword{datasets}
