% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WB_countries}
\alias{WB_countries}
\title{Countries and regions for World Bank datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 263 rows and 5 columns.
}
\source{
\url{https://data.worldbank.org/}
}
\usage{
WB_countries
}
\description{
A list of countries and regions as in World Bank Data datasets

The table is taken from the metadata of several datasets of the World Bank Data open data bank

\describe{
\item{country_code}{ISO 3166-1 alpha-3 country codes or regions}
\item{country_name}{The name of the country or region, as of the World Bank datasets}
\item{region}{The geograpical region of each country}
\item{income_group}{The income group of each country}
\item{notes}{Notes for specific observations}}
}
\examples{

WB_countries

}
\keyword{datasets}
