% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WB_fertility}
\alias{WB_fertility}
\title{Fertility rates of countries from World Bank Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14262 rows and 4 columns.
}
\source{
\url{https://data.worldbank.org/indicator/SP.DYN.TFRT.IN}
}
\usage{
WB_fertility
}
\description{
A table of fertility rates for each country and year

The total fertility rate (TFR), sometimes also called the fertility rate, absolute/potential natality, period total fertility rate (PTFR), or total period fertility rate (TPFR) of a population is the average number of children that would be born to a woman over her lifetime if she was to experience the exact current age-specific fertility rates through her lifetime and she was to live from birth until the end of her reproductive life.

\describe{
\item{country_code}{ISO 3166-1 alpha-3 country codes or regions}
\item{indicator}{The World Bank indicator (SP.DYN.TFRT.IN)}
\item{year}{The year of the datum}
\item{value}{The value of the datum}}
}
\examples{

WB_fertility

}
\keyword{datasets}
