% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WB_gdp_per_capita}
\alias{WB_gdp_per_capita}
\title{GDP per capita at purchasing power parity of countries from World Bank Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6961 rows and 4 columns.
}
\source{
\url{https://data.worldbank.org/indicator/NY.GDP.PCAP.PP.CD}
}
\usage{
WB_gdp_per_capita
}
\description{
A table of fertility GDP per capita for each country and year, in current international dollars

The purchasing power parity (PPP) value of all final goods and services produced within a country in a given year, divided by the average (or mid-year) population for the same year.

\describe{
\item{country_code}{ISO 3166-1 alpha-3 country codes or regions}
\item{indicator}{The World Bank indicator (NY.GDP.PCAP.PP.CD)}
\item{year}{The year of the datum}
\item{value}{The value of the datum}}
}
\examples{

WB_gdp_per_capita

}
\keyword{datasets}
