% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WB_gini}
\alias{WB_gini}
\title{Gini index of countries from World Bank Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1745 rows and 4 columns.
}
\source{
\url{https://data.worldbank.org/indicator/SI.POV.GINI}
}
\usage{
WB_gini
}
\description{
A table of the Gini index for each country and year, estimated by the World Bank

The Gini index or coefficient of inequality of income. A Gini index near one expresses maximal inequality, and a value near zero expresses maximal equality.

\describe{
\item{country_code}{ISO 3166-1 alpha-3 country codes or regions}
\item{indicator}{The World Bank indicator (SI.POV.GINI)}
\item{year}{The year of the datum}
\item{value}{The value of the datum}}
}
\examples{

WB_gini

}
\keyword{datasets}
