% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WB_population_female}
\alias{WB_population_female}
\title{Population (female)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14329 rows and 4 columns.
}
\source{
\url{https://data.worldbank.org/indicator/SP.POP.TOTL.FE.IN}
}
\usage{
WB_population_female
}
\description{
A table of female population for each country and year, estimated by the World Bank

\describe{
\item{country_code}{ISO 3166-1 alpha-3 country codes or regions}
\item{indicator}{The World Bank indicator (SP.POP.TOTL.FE.IN)}
\item{year}{The year of the datum}
\item{value}{The value of the datum}}
}
\examples{

WB_population_female

}
\keyword{datasets}
