% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WB_poverty_odd}
\alias{WB_poverty_odd}
\title{Poverty headcount ratio at 1.90 dollar a day}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2207 rows and 4 columns.
}
\source{
\url{https://data.worldbank.org/indicator/SI.POV.DDAY}
}
\usage{
WB_poverty_odd
}
\description{
A table of percentage of population earning less than 1.90 dollar a day (2011 purchasing power parity) for each country and year, estimated by the World Bank

\describe{
\item{country_code}{ISO 3166-1 alpha-3 country codes or regions}
\item{indicator}{The World Bank indicator (SI.POV.DDAY)}
\item{year}{The year of the datum}
\item{value}{The value of the datum}}
}
\examples{

WB_poverty_odd

}
\keyword{datasets}
