% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WorldMap1_10}
\alias{WorldMap1_10}
\title{A map of the world (1:10m).}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 255 rows and 90 columns.
}
\source{
\url{https://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-0-countries/}
}
\usage{
WorldMap1_10
}
\description{
A map of the world loaded as a sf object (scale 1:10m)

Countries distinguish between metropolitan (homeland) and independent and semi-independent portions of sovereign states. Each country is coded with a world region that roughly follows the United Nations setup. Countries are coded with standard ISO and FIPS codes. French INSEE codes are also included. Includes some thematic data from the United Nations (1), U.S. Central Intelligence Agency, and elsewhere. The map is in a 1:10m scale, suitable for detailed maps of specific regions.
}
\examples{

WorldMap1_10

}
\keyword{datasets}
