% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WorldMap1_110}
\alias{WorldMap1_110}
\title{A map of the world (1:110m).}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 177 rows and 87 columns.
}
\source{
\url{https://www.naturalearthdata.com/downloads/10m-cultural-vectors/110m-admin-0-countries/}
}
\usage{
WorldMap1_110
}
\description{
A map of the world loaded as a sf object (scale 1:110m)

Countries distinguish between metropolitan (homeland) and independent and semi-independent portions of sovereign states. Each country is coded with a world region that roughly follows the United Nations setup. Countries are coded with standard ISO and FIPS codes. French INSEE codes are also included. Includes some thematic data from the United Nations (1), U.S. Central Intelligence Agency, and elsewhere. The map is in a 1:110m scale, suitable for schematic maps of the world.
}
\examples{

WorldMap1_110

}
\keyword{datasets}
