% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WorldMapData}
\alias{WorldMapData}
\title{Data of world countries.}
\format{
An object of class \code{data.frame} with 177 rows and 12 columns.
}
\source{
data retrieved from the world bank, available in the spData package.
}
\usage{
WorldMapData
}
\description{
A data frame containing information from 188 world countries.

\describe{
\item{iso_a2}{ISO 3166-1 alpha-2 country codes.}
\item{name_long}{Long name of the country.}
\item{continent}{Continent the country belongs to.}
\item{region_un}{UN region.}
\item{subregion}{UN subregion.}
\item{type}{Variable indicating if it is a Sovereign Country, a Country, a Dependency, a Disputed territory or with Indeterminate status.}
\item{area_km2}{Country area.}
\item{pop}{Population.}
\item{liefExp}{Life expectancy.}
\item{gdpPercap}{gdp per capita.}
\item{lat}{Average latitude.}
\item{lon}{Average longitude.}
}
}
\examples{

WorldMapData

}
\keyword{datasets}
