% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{aq_bcn}
\alias{aq_bcn}
\title{Air quality data from Barcelona}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2225399 rows and 8 columns.
}
\source{
BCN Open Data \url{https://opendata-ajuntament.barcelona.cat/data/en/dataset/qualitat-aire-detall-bcn}
}
\usage{
aq_bcn
}
\description{
A dataset with data from air quality in Barcelona. Currently data from January 2020 to June 2024.

\describe{
\item{estacio}{Code of air quality measurement station. See aq_stations_bcn for more information.}
\item{codi_contaminant}{Code of pollutant. See aq_pollutants_bcn for more information.}
\item{any}{Year of data retrieval.}
\item{mes}{Month of data retrieval.}
\item{dia}{Day of data retrieval.}
\item{hora}{Hour of data retrieval.}
\item{value}{Value of pollutant in specified units (see aq_pollutants_bcn for more information).}
\item{data}{Date time in POSIXct format.}
}
}
\examples{

aq_bcn

}
\keyword{datasets}
