% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{BCNDistricts}
\alias{BCNDistricts}
\title{A map of barcelona districts}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 13 columns.
}
\source{
\url{https://vangdata.carto.com/tables/shapefiles_barcelona_distrito/public}
}
\usage{
BCNDistricts
}
\description{
A map of barcelona districts loaded as a sf object

District names have been set into utf8 encoding.
}
\examples{

BCNDistricts

}
\keyword{datasets}
