% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{esp_ccaa}
\alias{esp_ccaa}
\title{A map of Spanish Autonomous Regions.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 18 rows and 20 columns.
}
\usage{
esp_ccaa
}
\description{
A map with the borders of the Spanish autonomous regions as a sf object
}
\examples{

esp_ccaa

}
\keyword{datasets}
