% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{trs_cens}
\alias{trs_cens}
\title{A map of Terrassa census tracts}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 141 rows and 10 columns.
}
\source{
\url{https://emap.terrassa.cat/mapserver/zipService/downloadZIP.aspx?type=divadm&key=post&srs=ETRS89}
}
\usage{
trs_cens
}
\description{
A map of Terrassa census tracts loaded as a sf object

Additional information: AREA and PERIMETER of each zip code, district at DIST and census tract at SECC.
}
\examples{

trs_cens

}
\keyword{datasets}
