% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{trs_poligons}
\alias{trs_poligons}
\title{A map of Terrassa industrial parks}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 19 rows and 7 columns.
}
\source{
\url{https://opendata.terrassa.cat/es/dataset/mapa-dels-poligons-industrials}
}
\usage{
trs_poligons
}
\description{
A map of Terrassa industrial parks loaded as a sf object

Additional information: AREA in square meters, area in HECTARES and PERIMETER of each industrial park, and industrial park name in NOM_POLIGO.
}
\examples{

trs_poligons

}
\keyword{datasets}
