% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{USStates1_110}
\alias{USStates1_110}
\title{A map of states of the United States of America (scale 1:110m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 51 rows and 84 columns.
}
\source{
\url{https://www.naturalearthdata.com/downloads/110m-cultural-vectors/110m-admin-1-states-provinces/}
}
\usage{
USStates1_110
}
\description{
A map of states of the United States of America loaded as a sf object (scale 1:110m)

Internal, first-order administrative boundaries and polygons for the United States of America.
}
\examples{

USStates1_110

}
\keyword{datasets}
