% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{WorldStates1_10}
\alias{WorldStates1_10}
\title{A map of world states (scale 1:10m)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 4594 rows and 84 columns.
}
\source{
\url{https://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-1-states-provinces/}
}
\usage{
WorldStates1_10
}
\description{
A map of world states loaded as a sf object(scale 1:10m)

Internal, first-order administrative boundaries and polygons for all but a few tiny countries. Includes name attributes (including diacritical marks), name variants, and some statistical codes (FIPS, ISO, HASC).
}
\examples{

WorldStates1_10

}
\keyword{datasets}
